Add generic Linux installation target
The intent is to streamline downstream packaging and use to match more
"traditional" Linux environment.

This is expected to work across other *nix platforms with minimal or no
changes. All require() were adjusted to be absolute. The examples/ will
now work after installation is done or, for development purposes, with
PATH=$PWD and LUA_PATH="$PWD/?.lua;;"

CMake template files are not expected to work at this point, but they
were not working at all for me.